initial implementation of Models API
Initial implementation of Models part the Hagring API
(hapi).

We are using swagger 2.0 specification together with
connexion model that generates hooks automatically
from the specification.

Connexion is an extension to flask as web framework.
We also use SQLalchemy to access the sqlite database.

This package can be deployed as WSGI module behind
for example apache server.