check explicitly for model instances
Instead of relying on database constrains that prevent
a model from being deleted when it's instanciated on a
device, explicitly make a check if there are instances.

We need to do this way, so that it possible to remove
model's files from the S3.

Before removing files from S3, we must check if we
are allowed to do it.